\name{MANOR-internal}
\title{Internal Functions for MANOR Package}
\alias{MANOR-internal}
\alias{add.lines}
\alias{arrayCGH2txt}
\alias{center}
\alias{check.names}
\alias{chrom}
\alias{cv}
\alias{flag.aggregate.arrayCGH}
\alias{my.aggregate.data.frame}
\alias{print.flag}
\alias{rename}
\alias{w}
\alias{num.chromosome.flag}
\alias{replicate.aux}
\alias{spot.corr.flag}
\alias{getChromosomeArm}
\alias{intensity.plot}

\description{}

\author{Pierre Neuvial, \email{manor@curie.fr}.}

\note{People interested in tools for array-CGH analysis can
  visit our web-page: \url{http://bioinfo.curie.fr}.}

\keyword{internal}



